+++ KosmicTask/KosmicTask/source/MGSMinimalViewToolViewController.m
@@ -72,7 +72,7 @@
  segment click
  
  */
-- (void)segmentClick:(int)selectedSegment
+- (void)segmentClick:(NSInteger)selectedSegment
 {
 	eMGSMotherViewConfig viewConfig;
 	eMGSViewState viewState = kMGSViewStateNormalSize;
