+++ KosmicTask/KosmicTask/source/MGSResultToolViewController.h
@@ -19,7 +19,7 @@
 @property (readonly) MGSPopupButton *actionPopupButton;
 
 - (IBAction)segControlClicked:(id)sender;
-- (void)segmentClick:(int)selectedSegment;
+- (void)segmentClick:(NSInteger)selectedSegment;
 - (void)initialiseForWindow:(NSWindow *)window;
 
 @end
